refactor(auth): 인증 도메인 파일 위치 정리
