+++ src/pages/auth/components/authSharedStyles.ts
